pyproject.toml
[project]
name = "lingokeun-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-genai>=1.0.0",
    "pydantic-settings>=2.12.0",
    "typer>=0.21.1",
]

[project.scripts]
lingokeun = "lingokeun.main:generate"

[tool.uv]
package = true

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "pytest>=9.0.2",
    "ruff>=0.14.14",
]
